+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Improve missing `Content-Type` header error message
+
 1.0.0 / 2015-02-01
 ==================